requirements.txt
aiohttp==3.8.4
docx2pdf==0.1.8
Flask==2.3.2
flask_restx==1.1.0
langchain==0.0.271
openai==0.27.8
pydantic==1.10.12
pydantic==1.10.9
python-dotenv==1.0.0
python_docx==0.8.11
tiktoken==0.3.3
tinydb==4.8.0
lark
chroma
